+++ .idea/gradle.xml
@@ -10,6 +10,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/architecture" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
