--- app/src/main/java/com/sedooj/resumen/domain/data/user/create/CreateUserInput.kt
@@ -1,6 +1,0 @@
-package com.sedooj.resumen.domain.data.user.create
-
-data class CreateUserInput(
-    var username: String,
-    var password: String,
-)
