--- app/src/main/java/com/sedooj/resumen/domain/data/user/create/CreateUserOutput.kt
@@ -1,8 +1,0 @@
-package com.sedooj.resumen.domain.data.user.create
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CreateUserOutput(
-    var userId: Long
-)
