--- app/src/main/java/com/sedooj/resumen/domain/data/user/create/auth/AuthUserInput.kt
@@ -1,9 +1,0 @@
-package com.sedooj.resumen.domain.data.user.create.auth
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AuthUserInput(
-    var username: String = "",
-    var password: String = ""
-)
