--- app/src/main/java/com/sedooj/resumen/viewmodel/models/AuthenticationModel.kt
@@ -1,30 +1,0 @@
-package com.sedooj.resumen.viewmodel.models
-
-import android.content.Context
-import com.sedooj.resumen.domain.repository.user.UsersNetworkRepository
-import kotlinx.coroutines.CoroutineScope
-
-interface AuthenticationModel {
-
-    enum class AuthState {
-        NOT_AUTHORIZED,
-        AUTHORIZATION,
-        AUTHORIZED
-    }
-
-    fun updatePageState(state: AuthState)
-
-    fun resetErrorState()
-
-    fun setError(msg: Int)
-
-    fun validateInput(input: AuthorizationInput): Boolean
-
-    fun auth(
-        input: AuthorizationInput,
-        usersNetworkRepository: UsersNetworkRepository,
-        scope: CoroutineScope,
-        context: Context,
-    )
-
-}
