--- app/src/main/java/com/sedooj/resumen/viewmodel/models/AuthorizationInput.kt
@@ -1,6 +1,0 @@
-package com.sedooj.resumen.viewmodel.models
-
-data class AuthorizationInput(
-    var username: String,
-    var password: String
-)
