+++ architecture/src/main/java/com/sedooj/architecture/domain/data/user/create/CreateUserInput.kt
@@ -1,0 +1,6 @@
+package com.sedooj.architecture.domain.data.user.create
+
+data class CreateUserInput(
+    var username: String,
+    var password: String,
+)
