+++ architecture/src/main/java/com/sedooj/architecture/domain/data/user/create/CreateUserOutput.kt
@@ -1,0 +1,8 @@
+package com.sedooj.architecture.domain.data.user.create
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CreateUserOutput(
+    var userId: Long
+)
