+++ architecture/src/main/java/com/sedooj/architecture/domain/data/user/create/auth/AuthUserInput.kt
@@ -1,0 +1,9 @@
+package com.sedooj.architecture.domain.data.user.create.auth
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AuthUserInput(
+    var username: String = "",
+    var password: String = ""
+)
