+++ architecture/src/main/java/com/sedooj/architecture/storage/entity/AuthUserEntity.kt
@@ -1,0 +1,12 @@
+package com.sedooj.architecture.storage.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class AuthUserEntity(
+    @PrimaryKey val id: Int,
+    @ColumnInfo(name = "username") val username: String,
+    @ColumnInfo(name = "password") val password: String,
+)
