+++ architecture/src/main/java/com/sedooj/architecture/viewmodel/auth/model/AuthenticationModel.kt
@@ -1,0 +1,30 @@
+package com.sedooj.architecture.viewmodel.auth.model
+
+import android.content.Context
+import com.sedooj.resumen.domain.repository.user.UsersNetworkRepository
+import kotlinx.coroutines.CoroutineScope
+
+interface AuthenticationModel {
+
+    enum class AuthState {
+        NOT_AUTHORIZED,
+        AUTHORIZATION,
+        AUTHORIZED
+    }
+
+    fun updatePageState(state: AuthState)
+
+    fun resetErrorState()
+
+    fun setError(msg: Int)
+
+    fun validateInput(input: AuthorizationInput): Boolean
+
+    fun auth(
+        input: AuthorizationInput,
+        usersNetworkRepository: UsersNetworkRepository,
+        scope: CoroutineScope,
+        context: Context,
+    )
+
+}
