+++ architecture/src/main/java/com/sedooj/architecture/viewmodel/auth/model/AuthorizationInput.kt
@@ -1,0 +1,6 @@
+package com.sedooj.architecture.viewmodel.auth.model
+
+data class AuthorizationInput(
+    var username: String,
+    var password: String
+)
