+++ build.gradle.kts
@@ -3,5 +3,6 @@
 plugins {
     alias(libs.plugins.androidApplication) apply false
     alias(libs.plugins.kotlinAndroid) apply false
+    alias(libs.plugins.androidLibrary) apply false
 }
 true // Needed to make the Suppress annotation work for the plugins block
